# Fix edge case when query value siphash is zero