This PR reduces the number of **false positives** we get when querying the filters. It is an edge case but after querying thousands of filters with hundreds of keys the number of blocks downloaded is now much less.

It was heavily tested for correctness but it is not optimized (not measured yet).