@lontivero 2019-12-04
Yes, I did it as a draft just to prevent you merge it before being sure the performance is not worse.

---------------

The new algorithm performs almost exactly as the previous one.

### Current master

```
|           Method |      N |          Mean |      Error |     StdDev | Rank |
|----------------- |------- |--------------:|-----------:|-----------:|-----:|
| MatchAnySmallSet |   1000 |      5.332 us |  0.0664 us |  0.0519 us |    1 |
| MatchAnyBigSet   |   1000 |     62.704 us |  0.3804 us |  0.3372 us |    3 |
| MatchAnySmallSet |  10000 |     11.143 us |  0.1419 us |  0.1258 us |    2 |
| MatchAnyBigSet   |  10000 |    906.241 us |  9.9615 us |  8.8306 us |    5 |
| MatchAnySmallSet | 100000 |    224.626 us |  0.9653 us |  0.7537 us |    4 |
| MatchAnyBigSet   | 100000 | 11,534.003 us | 70.3391 us | 58.7363 us |    6 |
```

### This PR

```
|           Method |      N |          Mean |      Error |     StdDev | Rank |
|----------------- |------- |--------------:|-----------:|-----------:|-----:|
| MatchAnySmallSet |   1000 |      5.411 us |  0.0169 us |  0.0141 us |    1 |
| MatchAnyBigSet   |   1000 |     62.688 us |  0.4372 us |  0.3651 us |    3 |
| MatchAnySmallSet |  10000 |     11.083 us |  0.0614 us |  0.0544 us |    2 |
| MatchAnyBigSet   |  10000 |    902.267 us |  7.1631 us |  5.9815 us |    5 |
| MatchAnySmallSet | 100000 |    225.970 us |  2.1514 us |  1.7965 us |    4 |
| MatchAnyBigSet   | 100000 | 11,698.425 us | 59.7123 us | 52.9334 us |    6 |
```