@NicolasDorier 2019-12-05
does not build on AppVeyor.