@nopara73 2019-12-05
Maybe this one? `var key = blockHash[0..16];`