@yahiheb 2019-12-05
Or maybe `appveyor.yml` should be updated?